illumination/requirements.txt
gpiozero==2.0.1
lgpio==0.2.2.0
schedule==1.2.2
paho-mqtt==2.1.0
pydantic==2.7.4
